requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.0.4
ecdsa==0.17.0
fastapi==0.75.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
passlib==1.7.4
pyasn1==0.4.8
pydantic==1.9.0
PyMySQL==1.0.2
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
typing-extensions==4.1.1
uvicorn==0.17.6
zipp==3.7.0
